+++ modules/file_organization_module/file_organization.py
@@ -193,7 +193,11 @@ class FileOrganization(QWidget):
         # Documentation
         groupbox = QGroupBox("Documentation")
         layout2 = QVBoxLayout()
-        layout2.addWidget(label_documentation)
+        collapsible_widget = gf.CollapsibleWidget('', collapsed_icon="▶ (show)", expanded_icon="▼ (hide)", expanded=False)
+        collapsible_widget.content.setLayout(QVBoxLayout())
+        collapsible_widget.content.layout().addWidget(label_documentation)
+        collapsible_widget.content.layout().setContentsMargins(0,0,0,0)
+        layout2.addWidget(collapsible_widget)
         groupbox.setLayout(layout2)
         layout.addWidget(groupbox)