+++ modules/graph_filtering_module/graph_filtering.py
@@ -25,7 +25,11 @@ class GraphFiltering(QWidget):
                                     'Input segmentation mask must have X, Y and T axes. The optional input image must have X, Y and T axes and can optionally have C and/or Z axes.')
         groupbox = QGroupBox("Documentation")
         layout2 = QVBoxLayout()
-        layout2.addWidget(label_documentation)
+        collapsible_widget = gf.CollapsibleWidget('', collapsed_icon="▶ (show)", expanded_icon="▼ (hide)", expanded=False)
+        collapsible_widget.content.setLayout(QVBoxLayout())
+        collapsible_widget.content.layout().addWidget(label_documentation)
+        collapsible_widget.content.layout().setContentsMargins(0,0,0,0)
+        layout2.addWidget(collapsible_widget)
         groupbox.setLayout(layout2)
         layout.addWidget(groupbox)